+++ packages/gamut/src/Form/FormError.tsx
@@ -1,8 +1,17 @@
-import cx from 'classnames';
+import styled from '@emotion/styled';
 import React, { HTMLAttributes } from 'react';
 
-import styles from './styles/FormError.module.scss';
+import { colorStates } from './styles/shared';
+
+const ErrorSpan = styled.span`
+  position: absolute;
+  left: 0;
+  top: calc(100% - ${({ theme }) => theme.spacing[8]});
+  color: ${colorStates.error.color};
+  display: inline-block;
+  font-size: ${({ theme }) => theme.fontSize[14]};
+`;
 
 export const FormError: React.FC<HTMLAttributes<HTMLSpanElement>> = (props) => {
-  return <span className={cx(styles.formError, props.className)} {...props} />;
+  return <ErrorSpan {...props} />;
 };