+++ packages/gamut/src/GridForm/GridFormInputGroup/__fixtures__/renderers.tsx
@@ -1,6 +1,9 @@
-import { mount, ReactWrapper } from 'enzyme';
+import { theme } from '@codecademy/gamut-styles';
+import { ThemeProvider } from '@emotion/react';
+import { mount } from 'enzyme';
 import React from 'react';
 
+import { ToolTipProps } from '../../../ToolTip';
 import {
   stubCheckboxField,
   stubFileField,
@@ -23,11 +26,16 @@ import { GridFormRadioGroupInput } from '../GridFormRadioGroupInput';
 import { GridFormSelectInput } from '../GridFormSelectInput';
 import { GridFormTextArea } from '../GridFormTextArea';
 import { GridFormTextInput } from '../GridFormTextInput';
+import { GridFormToolTip } from '../GridFromToolTip';
+
+const mountWithTheme = (component: React.ReactNode) => {
+  return mount(<ThemeProvider theme={theme}>{component}</ThemeProvider>);
+};
 
 export const renderGridFormSelectInput = (
   extraProps: Partial<GridFormSelectField> = {}
-): ReactWrapper => {
-  return mount(
+) => {
+  return mountWithTheme(
     <GridFormSelectInput
       field={{ ...stubSelectField, ...extraProps }}
       register={jest.fn()}
@@ -39,7 +47,7 @@ export const renderGridFormSelectInput = (
 export const renderGridFormTextInput = (
   extraProps: Partial<GridFormTextField> = {}
 ) => {
-  return mount(
+  return mountWithTheme(
     <GridFormTextInput
       field={{ ...stubTextField, ...extraProps }}
       register={jest.fn()}
@@ -51,7 +59,7 @@ export const renderGridFormTextInput = (
 export const renderGridFormTextArea = (
   extraProps: Partial<GridFormTextAreaField> = {}
 ) => {
-  return mount(
+  return mountWithTheme(
     <GridFormTextArea
       field={{ ...stubTextareaField, ...extraProps }}
       register={jest.fn()}
@@ -63,7 +71,7 @@ export const renderGridFormTextArea = (
 export const renderGridFormRadioGroupInput = (
   extraProps: Partial<GridFormRadioGroupField> = {}
 ) => {
-  return mount(
+  return mountWithTheme(
     <GridFormRadioGroupInput
       field={{ ...stubRadioGroupField, ...extraProps }}
       setValue={jest.fn()}
@@ -75,8 +83,8 @@ export const renderGridFormRadioGroupInput = (
 
 export const renderGridFormFileInput = (
   extraProps: Partial<GridFormFileField> = {}
-): ReactWrapper => {
-  return mount(
+) => {
+  return mountWithTheme(
     <GridFormFileInput
       field={{ ...stubFileField, ...extraProps }}
       register={jest.fn()}
@@ -87,14 +95,18 @@ export const renderGridFormFileInput = (
 
 export const renderGridFormCheckboxInput = (
   extraProps: Partial<GridFormCheckboxField> = {}
-): ReactWrapper => {
-  return mount(
+) => {
+  return mountWithTheme(
     <GridFormCheckboxInput
       field={{ ...stubCheckboxField, ...extraProps }}
       register={jest.fn()}
       {...extraProps}
     />
   );
+};
+
+export const renderGridFormToolTip = (extraProps: ToolTipProps) => {
+  return mountWithTheme(<GridFormToolTip {...extraProps} />);
 };
 
 export const getComponent = (componentName: string, extraProps: any) => {
@@ -111,6 +123,8 @@ export const getComponent = (componentName: string, extraProps: any) => {
       return renderGridFormFileInput(extraProps);
     case 'GridFormCheckboxInput':
       return renderGridFormCheckboxInput(extraProps);
+    case 'GridFormToolTip':
+      return renderGridFormToolTip(extraProps);
     default:
       throw new Error(`Unknown component name: ${componentName}`);
   }