+++ packages/gamut/src/Modal/index.tsx
@@ -3,7 +3,7 @@ import cx from 'classnames';
 import React from 'react';
 
 import { ButtonDeprecated } from '../ButtonDeprecated';
-import { CardShell } from '../Card';
+import { CardShell } from '../DeprecatedCard';
 import { Overlay, OverlayProps } from '../Overlay';
 import styles from './styles.module.scss';
 
@@ -60,8 +60,8 @@ export const Modal: React.FC<ModalProps> = ({
         className={cx(styles.modalBody, className)}
         aria-hidden="false"
         aria-modal="true"
-        aria-label={ariaLabel}
         role="dialog"
+        aria-label={ariaLabel}
       >
         {!hideDefaultCloseButton && (
           <div