+++ packages/markdown-overrides/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.5.12](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.11...@codecademy/markdown-overrides@0.5.12) (2021-03-23)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
+### [0.5.11](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.10...@codecademy/markdown-overrides@0.5.11) (2021-03-19)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
+### [0.5.10](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.9...@codecademy/markdown-overrides@0.5.10) (2021-03-18)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
+### [0.5.9](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.8...@codecademy/markdown-overrides@0.5.9) (2021-03-16)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
 ### [0.5.8](https://github.com/Codecademy/client-modules/compare/@codecademy/markdown-overrides@0.5.7...@codecademy/markdown-overrides@0.5.8) (2021-03-04)
 
 **Note:** Version bump only for package @codecademy/markdown-overrides